backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.34.0
python-multipart>=0.0.20
openai>=1.30.0
Pillow>=11.0.0
pdf2image>=1.17.0
python-dotenv>=1.1.0
